Move github_mcp_oauth.py to integrations/github/mcp_oauth.py
