+++ surfaces/interactive_shell/runtime/startup/first_launch_github.py
@@ -96,7 +96,7 @@ def _print_intro(console: Console) -> None:
 
 
 def _show_device_code(console: Console, code: object) -> None:
-    from integrations.github_mcp_oauth import GitHubDeviceCode
+    from integrations.github.mcp_oauth import GitHubDeviceCode
 
     if not isinstance(code, GitHubDeviceCode):
         return
@@ -130,8 +130,9 @@ def _ask_retry(_console: Console) -> bool:
 
 def _attempt_login(console: Console) -> str:
     """Run one login attempt. Returns ``"success"``, ``"failed"``, or ``"quit"``."""
+
     from integrations.github.login import authenticate_and_configure_github
-    from integrations.github_mcp_oauth import GitHubDeviceFlowError
+    from integrations.github.mcp_oauth import GitHubDeviceFlowError
 
     try:
         result = authenticate_and_configure_github(
