+++ tests/integrations/test_github_login.py
@@ -4,9 +4,9 @@ import pytest
 
 from integrations.github import login as github_login
 from integrations.github.login import GitHubLoginResult
+from integrations.github.mcp_oauth import GitHubDeviceToken
 from integrations.github_identity import saved_github_username
 from integrations.github_mcp import GitHubMCPValidationResult
-from integrations.github_mcp_oauth import GitHubDeviceToken
 
 
 def test_authenticate_and_configure_github_happy_path(monkeypatch: pytest.MonkeyPatch) -> None:
